really prevent downstream commands while bit-banging
only check the command ready bit if it's unmasked:  a) the
handler is shared, so a timer, for instance, could cause us to
pick up a new command, even though masked, and b) keep the
command ready bit masked throughout the handler until we're sure
we want it enabled.  only enable it as we leave the handler.